refacto: move notes to folder features driven
