+++ src/notes/notes.module.ts
@@ -1,7 +1,7 @@
 import { MongooseModule } from '@nestjs/mongoose';
 import { NotesController } from './notes.controller';
 import { NotesService } from './notes.service';
-import { Note, NoteSchema } from './schemas/note.schema';
+import { Note, NoteSchema } from '../schemas/note.schema';
 import { Module } from '@nestjs/common';
 
 @Module({
