+++ src/notes/notes.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Note } from './schemas/note.schema';
+import { Note } from '../schemas/note.schema';
 import { Model } from 'mongoose';
-import { CreateNoteDto } from './dto/create-note.dto';
-import { UpdateNoteDto } from './dto/update-note.dto';
+import { CreateNoteDto } from '../dto/create-note.dto';
+import { UpdateNoteDto } from '../dto/update-note.dto';
 
 @Injectable()
 export class NotesService {
@@ -22,8 +22,15 @@ export class NotesService {
     return createdNote.save();
   }
 
-  async update(id: string, updateNoteDto: UpdateNoteDto): Promise<Note> {
-    return this.noteModel.findByIdAndUpdate(id, updateNoteDto);
+  async update(
+    author: string,
+    id: string,
+    updateNoteDto: UpdateNoteDto,
+  ): Promise<Note> {
+    return this.noteModel.findOneAndUpdate(
+      { _id: id, author: author },
+      updateNoteDto,
+    );
   }
 
   async destory(id: string) {
